updates seed file name and adds table truncations to tests
