+++ db/seeds/test/06-users.js
@@ -1,23 +1,16 @@
+const bcrypt = require('bcrypt')
+
+const password = bcrypt.hashSync("password", 10);
+
+
 exports.seed = (knex, Promise) => {
   return knex.raw('TRUNCATE users RESTART IDENTITY CASCADE')
   .then( () => {
     return Promise.all([
-      knex.raw(
-        'INSERT INTO users (name, email, password, token, role_id, created_at) VALUES (?, ?, ?, ?, ?, ?)',
-        ["Xena", "xena@xena.com", "password", "", 2, new Date]
-      ),
-      knex.raw(
-        'INSERT INTO users (name, email, password, token, role_id, created_at) VALUES (?, ?, ?, ?, ?, ?)',
-        ["Amanda Palmer", "afp@sademopunk.com", "password", "", 2, new Date]
-      ),
-      knex.raw(
-        'INSERT INTO users (name, email, password, token, role_id, created_at) VALUES (?, ?, ?, ?, ?, ?)',
-        ["Boudi the Cat", "perfection@imakitty.com", "password", "", 2, new Date]
-      ),
-      knex.raw(
-        'INSERT INTO users (name, email, password, token, role_id, created_at) VALUES (?, ?, ?, ?, ?, ?)',
-        ["admin", "admin@admin.com", "password", "", 1, new Date]
-      ),
+      knex('users').insert( { name: "Xena", email: "xena@xena.com", password: password, token: "", role_id: 2, created_at: new Date }),
+      knex('users').insert( { name: "Amanda Palmer", email: "afp@sademopunk.com", password: password, token: "", role_id: 2, created_at: new Date }),
+      knex('users').insert( { name: "Boudi the Kitty", email: "perfection@imakitty.com", password: password, token: "", role_id: 2, created_at: new Date }),
+      knex('users').insert( { name: "Admin", email: "admin@admin.com", password: password, token: "", role_id: 1, created_at: new Date }),
     ])
   })
 }
