+++ test/users-test.js
@@ -20,6 +20,25 @@ describe("Users", () => {
     })
   })
 
+  beforeEach(done => {
+    database.seed.run()
+    .then(() => done())
+  })
+
+  afterEach(done => {
+    Promise.all([
+      database.raw(`TRUNCATE lifts RESTART IDENTITY CASCADE`),
+      database.raw(`TRUNCATE bodyareas RESTART IDENTITY CASCADE`),
+      database.raw(`TRUNCATE bodyarea_lifts RESTART IDENTITY CASCADE`),
+      database.raw(`TRUNCATE workouts RESTART IDENTITY CASCADE`),
+      database.raw(`TRUNCATE workout_lifts RESTART IDENTITY CASCADE`),
+      database.raw(`TRUNCATE sets RESTART IDENTITY CASCADE`),
+      database.raw(`TRUNCATE roles RESTART IDENTITY CASCADE`),
+      database.raw(`TRUNCATE users RESTART IDENTITY CASCADE`),
+    ])
+    .then(() => done())
+  })
+
   after( () => {
     this.server.close()
   })
